+++ src/pr_sorter.cpp
@@ -19,12 +19,12 @@ void PrSorted::sort_links(std::vector<std::pair<std::string, int>>& linksVector,
         num_res = linksVector.size() -1;
     }
 
-    std::shared_ptr<std::unordered_map<std::string, float>> scores = m_score.getMap();
+    std::shared_ptr<std::unordered_map<std::string, double>> scores = m_score.getMap();
       
     std::partial_sort(linksVector.begin(), linksVector.begin() + num_res, linksVector.end(),
         [&scores](const std::pair<std::string, size_t>& part1, const std::pair<std::string, size_t>& part2){
-            return (static_cast<float>(part1.second) * scores->at(part1.first)) >
-            (static_cast<float>(part2.second) * scores->at(part2.first));
+            return (static_cast<double>(part1.second) * scores->at(part1.first)) >
+            (static_cast<double>(part2.second) * scores->at(part2.first));
         }
     );    
 }
